==> code/data_processing/01_API_NBA.ipynb <==
<div class="alert alert-danger">
    <h4 style="font-weight: bold; font-size: 28px;">NBA API</h4>
    <p style="font-size: 20px;">Data Gathering</p>
</div>

<a name="NBA"></a>

# Setup

import pandas as pd
from datetime import datetime, timedelta
import time

from nba_api.stats.static import teams, players
from nba_api.stats.endpoints import (
  scoreboard, leaguegamefinder, playercareerstats,
  boxscorematchupsv3, boxscoreadvancedv2, teamestimatedmetrics, 
  teamgamelogs, TeamGameLogs, TeamEstimatedMetrics, leaguedashteamstats,
    hustlestatsboxscore, boxscoremiscv2, boxscorefourfactorsv2,
    boxscorescoringv2, boxscoreusagev2, boxscoreplayertrackv2
)

# Team Data

# get_teams returns a list of 30 dictionaries, each an NBA team
nba_teams = teams.get_teams()
print("Number of teams fetched: {}".format(len(nba_teams)))
nba_teams_df = pd.DataFrame(nba_teams)
nba_teams_df.head()

# Players

# get_players returns a list of dictionaries, each representing a player
nba_players = players.get_players()
print("Number of players fetched: {}".format(len(nba_players)))
nba_players_df = pd.DataFrame(nba_players)
nba_players_df.head()

# Scoreboard

# Today's Score Board
games = scoreboard.ScoreBoard()

# json
games.get_json()

# dictionary
games.get_dict()

# League Game Finder

# get game data
team_ids = nba_teams_df['id'].tolist()

games_list = []

for id in team_ids:
    print(id)
    # query for games
    gamefinder = leaguegamefinder.LeagueGameFinder(team_id_nullable=id)
    # we want the first DataFrame of those returned
    games_list.append(gamefinder.get_data_frames()[0])
    # add time delay between requests
    time.sleep(3)

games_df = pd.concat(games_list)
print(games_df.shape)
games_df.head()

games_df['GAME_DATE'] = pd.to_datetime(games_df['GAME_DATE'])

# get the latest date
earliest_date = games_df['GAME_DATE'].max()
print(earliest_date)

games_df.to_csv('../data/original/nba_games_box_scores_1984_2024.csv', index=False)

# League Dash Team Stats

ldts = leaguedashteamstats.LeagueDashTeamStats(month=5, season='2023-24')

ldts.get_data_frames()[0].head()

ldts_list = []
seasons = []
for year in range(1996, 2024):
    season = f"{year}-{str(year + 1)[-2:]}"
    seasons.append(season)
months = range(1, 13)

for season in seasons:
    for month in months:
        print(f"Querying season {season}, month {month}")
        # query for months
        ldts = leaguedashteamstats.LeagueDashTeamStats(month=month, season=season)
        # get the first DataFrame of those returned
        df = ldts.get_data_frames()[0]
        
        # add columns for 'season' and 'month'
        df['SEASON'] = season
        df['MONTH'] = month
        
        # append the DataFrame to the list
        ldts_list.append(df)
        
        # add time delay between requests
        time.sleep(3)

# concatenate all DataFrames in the list into one large DataFrame
ldts_df = pd.concat(ldts_list, ignore_index=True)

ldts_df.head()

ldts_df.to_csv('../data/original/nba_dash_team_stats_1997_2024.csv', index=False)

ldts_df.value_counts('TEAM_NAME')

BC_2023 = ldts_df[(ldts_df['TEAM_NAME'] == 'Boston Celtics') & (ldts_df['SEASON'] == '2022-23')]
BC_2023.head()

# Player Career Statistics

# get player data
player_ids = nba_players_df['id'].tolist()

players_stats_list = []

for id in player_ids:
    print(id)
    # query for games
    career = playercareerstats.PlayerCareerStats(player_id=id)
    # we want the first DataFrame of those returned
    players_stats_list.append(career.get_data_frames()[0])
    # add time delay between requests
    time.sleep(1)

players_stats_df = pd.concat(players_stats_list)
print(players_stats_df.shape)
players_stats_df.head()

# get the latest season
earliest_date = players_stats_df['SEASON_ID'].min()
print(earliest_date)

players_stats_df.to_csv('../data/original/nba_players_statistics_1946_2024.csv', index=False)

# Box Score Matchups V3

games_df = pd.read_csv('../data/original/nba_games_box_scores_1984_2024.csv')
# get player data
game_ids = games_df['GAME_ID'].tolist()

game_ids[5000]

bs_matchups = boxscorematchupsv3.BoxScoreMatchupsV3(game_id=21000400)
bs_matchups.get_data_frames()[0]

# Box Score Advanced V3

bs_adv = boxscoreadvancedv2.BoxScoreAdvancedV2(game_id=21000400)
check = bs_adv.get_data_frames()[1]
check

# function to get game_ids
def get_game_ids(season_id):
    game_ids = []
    game_ids = games_df['GAME_ID'][(games_df.SEASON_ID == season_id)].tolist()
    return game_ids

#get game ids for 2023 - 2024 season
game_ids_2023_2024 = get_game_ids('22023')

#get game ids for 2022 - 2023 season
game_ids_2022_2023 = get_game_ids('22022')

#get game ids for 2021 - 2022 season
game_ids_2021_2022 = get_game_ids('22021')

# function to get team advanced stats per game for a given season
def get_adv_stats_df(game_id_list):
    adv_games_stats_list = []
    for id in game_id_list:
        print(id)
        # query for games
        games = boxscoreadvancedv3.BoxScoreAdvancedV3(game_id=id)
        adv_games_stats_list.append(games.get_data_frames()[1])
        time.sleep(3)
    adv_stats_df = pd.concat(adv_games_stats_list, ignore_index=True)
    adv_stats_df = adv_stats_df.drop_duplicates()
    return adv_stats_df

# get advanced stats df for 2021 - 2022 season
adv_stats_df_2021_2022 = get_adv_stats_df(game_ids_2021_2022)
adv_stats_df_2021_2022.head()

# get advanced stats df for 2022 - 2023 season
adv_stats_df_2022_2023 = get_adv_stats_df(game_ids_2022_2023)
adv_stats_df_2022_2023.head()

# get advanced stats df for 2023 - 2024 season
adv_stats_df_2023_2024 = get_adv_stats_df(game_ids_2023_2024)
adv_stats_df_2023_2024.head()

#combine advanced stat dataframes into one combined datafram
adv_stats_frames = [adv_stats_df_2021_2022, adv_stats_df_2022_2023, adv_stats_df_2023_2024]
adv_stats_df = pd.concat(adv_stat_frames)
adv_stats_df.head()

#creates team name field to match games_df
adv_stats_df['TEAM_NAME'] = adv_stats_df['teamCity'] + " " + adv_stats_df['teamName']

#renames fields that match games_df
adv_stats_df.rename(columns={'gameId':'GAME_ID','teamId':'TEAM_ID', 'teamTricode': 'TEAM_ABBREVIATION'}, inplace=True)

#drop redundant columns
adv_stats_df.drop(['teamCity', 'teamName', 'teamSlug'], inplace=True, axis=1)

adv_stats_df.head()

#get GAME_DATE, MATCHUP, GAME_ID, TEAM_ABBREVIATION fields from games_df
adv_stats_df = pd.merge(adv_stats_df, games_df[['SEASON_ID','GAME_DATE','MATCHUP', 'GAME_ID','TEAM_ID', 'TEAM_ABBREVIATION']], on=['GAME_ID','TEAM_ID', 'TEAM_ABBREVIATION'])

adv_stats_df.head()

#export csv
adv_stats_df.to_csv('../data/original/nba_advanced_statistics_2021_2024.csv', index=False)

# Team Game Logs

mav_id = '1610612742'
logs = teamgamelogs.TeamGameLogs.DataSet(data=)
#logs.get_data_frame(data=)

# Team Estimated Metrics

team_metrics = teamestimatedmetrics.TeamEstimatedMetrics.DataSet(data=)

from nba_api.stats.endpoints import commonplayerinfo

# Basic Request
player_info = commonplayerinfo.CommonPlayerInfo(player_id=2544)

player_info.available_seasons.get_data_frame()

# Team Hustle Stats

hustle = hustlestatsboxscore.HustleStatsBoxScore(game_id = "0022100209")
hustle.get_data_frames()[2]

# function to get team hustle stats per game for a given season
def get_hustle_stats_df(game_id_list):
    hustle_games_stats_list = []
    for id in game_id_list:
        print(id)
        # query for games
        hustle = hustlestatsboxscore.HustleStatsBoxScore(game_id=id)
        hustle_games_stats_list.append(hustle.get_data_frames()[2])
        time.sleep(1)
    hustle_stats_df = pd.concat(hustle_games_stats_list, ignore_index=True)
    hustle_stats_df = hustle_stats_df.drop_duplicates()
    return hustle_stats_df

# get hustle stats df for 2021 - 2022 season
#hustle_stats_df_2021_2022 = get_hustle_stats_df(game_ids_2021_2022)
hustle_stats_df_2021_2022.head()

# get hustle stats df for 2022 - 2023 season
#hustle_stats_df_2022_2023 = get_hustle_stats_df(game_ids_2022_2023)
hustle_stats_df_2022_2023.head()

# get hustle stats df for 2023 - 2024 season
#hustle_stats_df_2023_2024 = get_hustle_stats_df(game_ids_2023_2024)
hustle_stats_df_2023_2024.head()

#combine hustle stat dataframes into one combined dataframe
hustle_stats_frames = [hustle_stats_df_2021_2022, hustle_stats_df_2022_2023, hustle_stats_df_2023_2024]
hustle_stats_df = pd.concat(hustle_stats_frames)
hustle_stats_df.head()

#creates team name field to match games_df
hustle_stats_df['TEAM_NAME'] = hustle_stats_df['TEAM_CITY'] + " " + hustle_stats_df['TEAM_NAME']

#renames fields that match games_df
#hustle_stats_df.rename(columns={'gameId':'GAME_ID','teamId':'TEAM_ID', 'teamTricode': 'TEAM_ABBREVIATION'}, inplace=True)

#drop redundant columns
hustle_stats_df.drop(['TEAM_CITY'], inplace=True, axis=1)

hustle_stats_df.head()

hustle_stats_df = pd.merge(hustle_stats_df, games_df[['SEASON_ID','GAME_DATE','MATCHUP', 'GAME_ID','TEAM_ID', 'TEAM_ABBREVIATION']], on=['GAME_ID','TEAM_ID', 'TEAM_ABBREVIATION'])

hustle_stats_df.head()

#export csv
hustle_stats_df.to_csv('../data/original/nba_hustle_statistics_2021_2024.csv', index=False)

# Box Score Misc

misc = boxscoremiscv2.BoxScoreMiscV2(game_id = "0022100209")
misc.get_data_frames()[1]

# function to get team hustle stats per game for a given season
def get_misc_stats_df(game_id_list):
    misc_games_stats_list = []
    for id in game_id_list:
        print(id)
        # query for games
        misc = boxscoremiscv2.BoxScoreMiscV2(game_id=id)
        misc_games_stats_list.append(misc.get_data_frames()[1])
        time.sleep(1)
    misc_stats_df = pd.concat(misc_games_stats_list, ignore_index=True)
    misc_stats_df = misc_stats_df.drop_duplicates()
    return misc_stats_df

# get miscellaneous stats df for 2021 - 2022 season
misc_stats_df_2021_2022 = get_misc_stats_df(game_ids_2021_2022)
misc_stats_df_2021_2022.head()

# get miscellaneous stats df for 2022 - 2023 season
misc_stats_df_2022_2023 = get_misc_stats_df(game_ids_2022_2023)
misc_stats_df_2022_2023.head()

# get miscellaneous stats df for 2023 - 2024 season
misc_stats_df_2023_2024 = get_misc_stats_df(game_ids_2023_2024)
misc_stats_df_2023_2024.head()

#combine hustle stat dataframes into one combined dataframe
misc_stats_frames = [misc_stats_df_2021_2022, misc_stats_df_2022_2023, misc_stats_df_2023_2024]
misc_stats_df = pd.concat(misc_stats_frames)
misc_stats_df.head()

#creates team name field to match games_df
misc_stats_df['TEAM_NAME'] = misc_stats_df['TEAM_CITY'] + " " + misc_stats_df['TEAM_NAME']

#renames fields that match games_df
#hustle_stats_df.rename(columns={'gameId':'GAME_ID','teamId':'TEAM_ID', 'teamTricode': 'TEAM_ABBREVIATION'}, inplace=True)

#drop redundant columns
misc_stats_df.drop(['TEAM_CITY'], inplace=True, axis=1)

misc_stats_df.head()

misc_stats_df = pd.merge(misc_stats_df, games_df[['SEASON_ID','GAME_DATE','MATCHUP', 'GAME_ID','TEAM_ID', 'TEAM_ABBREVIATION']], on=['GAME_ID','TEAM_ID', 'TEAM_ABBREVIATION'])

misc_stats_df.head()

#export csv
misc_stats_df.to_csv('../data/original/nba_misc_boxscore_statistics_2021_2024.csv', index=False)

# Four Factors

test = boxscorefourfactorsv2.BoxScoreFourFactorsV2(game_id = "0022100209")
test.get_data_frames()[1]

# function to get team four factor stats per game for a given season
def get_factor_stats_df(game_id_list):
    factor_games_stats_list = []
    for id in game_id_list:
        print(id)
        # query for games
        games = boxscorefourfactorsv2.BoxScoreFourFactorsV2(game_id=id)
        factor_games_stats_list.append(games.get_data_frames()[1])
        #time.sleep(1)
    factor_stats_df = pd.concat(factor_games_stats_list, ignore_index=True)
    factor_stats_df = factor_stats_df.drop_duplicates()
    return factor_stats_df

# get four factor stats df for 2021 - 2022 season
#factor_stats_df_2021_2022 = get_factor_stats_df(game_ids_2021_2022)
factor_stats_df_2021_2022.head()

# get four factor stats df for 2022 - 2023 season
#factor_stats_df_2022_2023 = get_factor_stats_df(game_ids_2022_2023)
factor_stats_df_2022_2023.head()

# get four factor stats df for 2023 - 2024 season
#factor_stats_df_2023_2024 = get_factor_stats_df(game_ids_2023_2024)
factor_stats_df_2023_2024.head()

#combine hustle stat dataframes into one combined dataframe
factor_stats_frames = [factor_stats_df_2021_2022, factor_stats_df_2022_2023, factor_stats_df_2023_2024]
factor_stats_df = pd.concat(factor_stats_frames)
factor_stats_df.head()

import os

# Get the current working directory of the Python interpreter
cwd = os.getcwd()

# Print the current working directory
print(cwd)

#export csv
factor_stats_df.to_csv('../data/original/nba_four_factors_statistics_2021_2024.csv', index=False)

# Box Score Scoring

test = boxscorescoringv2.BoxScoreScoringV2(game_id = "0022100209")
test.get_data_frames()[1]

# function to get team four factor stats per game for a given season
def get_scoring_stats_df(game_id_list):
    scoring_games_stats_list = []
    for id in game_id_list:
        print(id)
        # query for games
        games = boxscorescoringv2.BoxScoreScoringV2(game_id=id)
        scoring_games_stats_list.append(games.get_data_frames()[1])
        #time.sleep(1)
    scoring_stats_df = pd.concat(scoring_games_stats_list, ignore_index=True)
    scoring_stats_df = scoring_stats_df.drop_duplicates()
    return scoring_stats_df

# get scoring stats df for 2021 - 2022 season
#scoring_stats_df_2021_2022 = get_scoring_stats_df(game_ids_2021_2022)
scoring_stats_df_2021_2022.head()

# get scoring stats df for 2022 - 2023 season
#scoring_stats_df_2022_2023 = get_scoring_stats_df(game_ids_2022_2023)
scoring_stats_df_2022_2023.head()

# get scoring stats df for 2023 - 2024 season
#scoring_stats_df_2023_2024 = get_scoring_stats_df(game_ids_2023_2024)
scoring_stats_df_2023_2024.head()

#combine scoring stat dataframes into one combined dataframe
scoring_stats_frames = [scoring_stats_df_2021_2022, scoring_stats_df_2022_2023, scoring_stats_df_2023_2024]
scoring_stats_df = pd.concat(scoring_stats_frames)
scoring_stats_df.head()

#export csv
scoring_stats_df.to_csv('../data/original/nba_scoring_statistics_2021_2024.csv', index=False)

# Box Score Player Track

test = boxscoreplayertrackv2.BoxScorePlayerTrackV2(game_id = "0022100209")
test.get_data_frames()[1]

# function to get team four factor stats per game for a given season
def get_track_stats_df(game_id_list):
    track_games_stats_list = []
    for id in game_id_list:
        print(id, len(track_games_stats_list))
        # query for games
        games = boxscoreplayertrackv2.BoxScorePlayerTrackV2(game_id=id)
        track_games_stats_list.append(games.get_data_frames()[1])
        #time.sleep(1)
    track_stats_df = pd.concat(track_games_stats_list, ignore_index=True)
    track_stats_df = track_stats_df.drop_duplicates()
    return track_stats_df

# get player track stats df for 2021 - 2022 season
#track_stats_df_2021_2022 = get_track_stats_df(game_ids_2021_2022)
track_stats_df_2021_2022.head()

# get player track stats df for 2021 - 2022 season
#track_stats_df_2022_2023 = get_track_stats_df(game_ids_2022_2023)
track_stats_df_2022_2023.head()

# get player track stats df for 2021 - 2022 season
#track_stats_df_2023_2024 = get_track_stats_df(game_ids_2023_2024)
track_stats_df_2023_2024.head()

#combine track stat dataframes into one combined dataframe
track_stats_frames = [track_stats_df_2021_2022, track_stats_df_2022_2023, track_stats_df_2023_2024]
track_stats_df = pd.concat(track_stats_frames)
track_stats_df.head()

#export csv
track_stats_df.to_csv('../data/original/nba_track_statistics_2021_2024.csv', index=False)
