==> code/data_processing/03i_Feature_Engineering_PlayerStats.ipynb <==
<div class="alert alert-danger">
    <h4 style="font-weight: bold; font-size: 28px;">Feature Engineering - Player Stats Per Game 2019-2024</h4>
    <p style="font-size: 20px;">NBA API Data (2022-2024)</p>
</div>

<a name="Feature-Engineering"></a>

# Table of Contents

[Setup](#Setup)

[Explanation](#Explanation)


[Step by Step To Pull Player Stats](#Step-by-Step-To-Pull-Player-Stats)


[Unused Player API Functionss](#Unused-Player-API-Functions)

# Setup

[Return to top](#Feature-Engineering)

import sys
from pathlib import Path
# get current working directory
cwd = %pwd
# add shared_code directory to Python sys.path
sys.path.append(str(Path(cwd).parent / "shared_code"))
# import all libraries in shared_code directory 'imports.py' file
from imports import *
%matplotlib inline

# import other libraries
import numpy as np
import time
from nba_api.stats.endpoints import playergamelog
from nba_api.stats.endpoints import playergamelogs

from nba_api.stats.endpoints import CommonAllPlayers

from nba_api.stats.endpoints import CommonPlayerInfo
from nba_api.stats.endpoints import CommonTeamRoster



# Pulling Player Stats

## Explanation

- I have tried several nba_api functions, but they generally do not work for pulling player stats 
- For instance, boxscoreplayertrackv2 is deprecated for 2021-24 seasons, cannot be pulled  
- CommonAllPlayers is deprecrated, returns only 100-120 players for seasons prior to this season
- playergamelog doesn't have Team ID
- CommonTeamRoster only shows ending team roster, but  players change teams throughout the season
- playercareerstats only has season stats, not per game


- To get player stats per game, for each season, we need to:
1. Loop through each team for each season, pull rosters from CommonTeamRosters
2. Aggregate and pull unique player_ids for each season for each roster
3. Then we loop through 'playergamelogs' (different function than playergamelog), and aggregate

## Step by Step To Pull Player Stats

[Return to top](#Feature-Engineering)

# We use nba game box scores to check later work
# Get total box score df
team_bs_df = pd.read_csv('../../data/original/nba_games_box_scores_2022_2024.csv')

# Get season and game IDs into list
# we check unique games later to make sure pull is legitimate
id_df = team_bs_df[['SEASON_ID', 'GAME_ID', 'TEAM_ID']]
id_list = id_df.values.tolist()

team_id_list = id_df['TEAM_ID'].unique()
len(team_id_list)
print(team_id_list)

# Example Player Roster    
player_roster = CommonTeamRoster(
    team_id= '1610612737',
    season = '2019-20', # change year(s) if needed
    league_id_nullable= '00' # nba 00, g_league 20, wnba 10
)

df_player_roster = player_roster.get_data_frames()[0]
df_player_roster

# specify seasons to pull
seasons = ['2019-20', '2020-21', '2021-22', '2022-23', '2023-24']

# declare lists
#player_rosters_list = []
#player_rosters_col_names = []
all_player_roster_df = pd.DataFrame()

# Loop Through Each Roster for each season    
for season in seasons:
    for team_id in team_id_list:
        player_roster = CommonTeamRoster(
            team_id= team_id,
            season = season, # change year(s) if needed
            league_id_nullable= '00' # nba 00, g_league 20, wnba 10
        )
        
        # get values into df
        df_player_roster = player_roster.get_data_frames()[0]
        
        # add column for season_year
        df_player_roster['SEASON_YEAR'] = season

        # concat to previous df
        all_player_roster_df = pd.concat([all_player_roster_df, df_player_roster], ignore_index=True)

all_player_roster_df.head()

# Check that playergamelogs is pulling correctly for single player
# because looping through all players and seasons will take 30 mins

# Initialize an empty DataFrame to store all game logs
check_all_seasons_logs_df = pd.DataFrame()

# List of seasons to loop through (update this list as needed)
# '2019-20', '2020-21', '2021-22', '2022-23', '2023-24'
seasons = ['2019-20', '2020-21', '2021-22', '2022-23', '2023-24']

# Fetch game logs for each season and add a 'SEASON' column
for season in seasons:
    player_logs = playergamelogs.PlayerGameLogs(player_id_nullable='203500', season_nullable= season, season_type_nullable= "Regular Season")
    season_logs_df = player_logs.get_data_frames()[0]
    check_all_seasons_logs_df = pd.concat([check_all_seasons_logs_df, season_logs_df], ignore_index=True)

check_all_seasons_logs_df.head()

##### Warning - this cell will take 30-40 minutes to run
# make sure code is running properly before running this cell

# Loop through all seasons for all players

# Initialize an empty DataFrame to store all game logs
all_seasons_logs_df = pd.DataFrame()

# List of seasons to loop through (update this list as needed)
seasons = ['2019-20', '2020-21', '2021-22', '2022-23', '2023-24']

for season in seasons:
    # Get unique ids in specific season
    unique_player_ids = all_player_roster_df[all_player_roster_df['SEASON_YEAR'] == season]['PLAYER_ID'].unique()
    
    # Loop through unique IDS
    for unique_id in unique_player_ids:
        # Get player logs
        player_logs = playergamelogs.PlayerGameLogs(player_id_nullable=unique_id, 
                                                    season_nullable= season,
                                                    season_type_nullable= "Regular Season")
        # get player game data into df
        season_logs_df = player_logs.get_data_frames()[0]
        # concat with master df
        all_seasons_logs_df = pd.concat([all_seasons_logs_df, season_logs_df], ignore_index=True)

len(all_seasons_logs_df)

# Check that each season has pulled the correct amount of games (should be 1230)
# 2019-20 will have less games becasue of covid
len(all_seasons_logs_df[all_seasons_logs_df['SEASON_YEAR'] == '2022-23']['GAME_ID'].unique())


all_seasons_logs_df.tail(20)

# Save file
all_seasons_logs_df.to_csv('../../data/original/nba_players_statistics_per_game_2019_2024.csv', index=False)

# Check nba_games_box_scores_2022-2024 for unique SEASON_IDs
id_df['SEASON_ID'].unique()

# check games with unique SEASON_IDs
len(id_df[id_df['SEASON_ID'] == 22022])

len(id_df['GAME_ID'].unique())

## Unused Player API Functions

## Doesn't work because commonAllplayers only pulls 100-120 players for 2019-2023 seasons

# specify seasons to pull
seasons = ['2019-20', '2020-21', '2021-22', '2022-23', '2023-24']

# loop through seasons
player_list = []
for season in seasons:
    common_all_players = CommonAllPlayers(
        is_only_current_season = 1, # 1 active, 0 not active
        league_id = '00', # nba 00, g_league 20, wnba 10
        season = season # change year(s) if needed
    )

    df_common_players = common_all_players.get_data_frames()[0]
    player_list.append(df_common_players.values.tolist())

len(pd.DataFrame(player_list[3]))

# Show example of player box scores for a given game id
player_boxscores = boxscoreplayertrackv2.BoxScorePlayerTrackV2(game_id = '0022200021')
df_player_boxscores = player_boxscores.get_data_frames()[0]
df_player_boxscores.head()


from nba_api.stats.endpoints import playercareerstats
# Fetching career statistics for Player of Choice using his player ID
player_career = playercareerstats.PlayerCareerStats(player_id='203500')
player_career_df = player_career.get_data_frames()[0]

# Extracting the seasons of player of choice
seasons_played = player_career_df['SEASON_ID'].unique()
print(seasons_played.tolist())

# Show example of player box scores for a given game id
player_boxscores = playergamelog.PlayerGameLog(player_id= '203925', season=2023)
df_player_boxscores = player_boxscores.get_data_frames()[0]
df_player_boxscores.head()

from nba_api.stats.endpoints import CommonTeamRoster

common_team_roster = CommonTeamRoster(
    team_id = '1610612752', # input team id
    league_id_nullable = '00', # nba 00, g_league 20, wnba 10
    season='2023-24')
df_common_team_roster = common_team_roster.get_data_frames()[0]
df_common_team_roster.head()
