==> code/data_processing/04a_Feature_Selection_Filter_Methods.ipynb <==
<div class="alert alert-danger">
    <h4 style="font-weight: bold; font-size: 28px;">Feature Selection using Filter Methods</h4>
    <p style="font-size: 20px;">NBA API Data (2022-2024)</p>
</div>

<a name="Feature-Selection"></a>

# Table of Contents

[Setup](#Setup)

[Data](#Data)

**[1. Filter Methods for Total Points](#1.-Filter-Methods-for-Total-Points)**

**[2. Filter Methods for Plus Minus](#2.-Filter-Methods-for-Plus-Minus)**

**[3. Filter Methods for Game Winner](#3.-Filter-Methods-for-Game-Winner)**

# Setup

[Return to top](#Feature-Selection)

import sys
from pathlib import Path
# get current working directory
cwd = %pwd
# add shared_code directory to Python sys.path
sys.path.append(str(Path(cwd).parent / "shared_code"))
# import all libraries in shared_code directory 'imports.py' file
from imports import *
%matplotlib inline

# Data

[Return to top](#Feature-Selection)

# load all the rolling data
boxscores = pd.read_csv('../../data/processed/nba_team_matchups_rolling_box_scores_2022_2024_r05.csv')
opponent = pd.read_csv('../../data/processed/nba_team_matchups_rolling_opponent_impact_box_scores_2022_2024_r05.csv')
advanced = pd.read_csv('../../data/processed/nba_team_matchups_rolling_adv_stats_2021_2024_r05.csv')
hustle = pd.read_csv('../../data/processed/nba_team_matchups_rolling_hustle_stats_2021_2024_r05.csv')
miscbox = pd.read_csv('../../data/processed/nba_team_matchups_rolling_misc_box_stats_2021_2024_r05.csv')
factor = pd.read_csv('../../data/processed/nba_team_matchups_rolling_four_factor_stats_2021_2024_r05.csv')
scoring = pd.read_csv('../../data/processed/nba_team_matchups_rolling_scoring_box_stats_2021_2024_r05.csv')
track = pd.read_csv('../../data/processed/nba_team_matchups_rolling_player_track_box_stats_2021_2024_r05.csv')

# merge all the features
cols_to_drop = ['GAME_RESULT', 'TOTAL_PTS', 'PLUS_MINUS', 'HOME_TEAM_NAME',
                'SEASON_ID', 'GAME_DATE', 'AWAY_TEAM_NAME', 'ROLL_HOME_PTS', 'ROLL_AWAY_PTS']

features_df = boxscores.merge(opponent.drop([col for col in cols_to_drop if col in opponent.columns], axis=1), how='left', on='GAME_ID')
features_df = features_df.merge(advanced.drop([col for col in cols_to_drop if col in advanced.columns], axis=1), how='left', on='GAME_ID')
features_df = features_df.merge(hustle.drop([col for col in cols_to_drop if col in hustle.columns], axis=1), how='left', on='GAME_ID')
features_df = features_df.merge(miscbox.drop([col for col in cols_to_drop if col in miscbox.columns], axis=1), how='left', on='GAME_ID')
features_df = features_df.merge(factor.drop([col for col in cols_to_drop if col in factor.columns], axis=1), how='left', on='GAME_ID')
features_df = features_df.merge(scoring.drop([col for col in cols_to_drop if col in scoring.columns], axis=1), how='left', on='GAME_ID')
features_df = features_df.merge(track.drop([col for col in cols_to_drop if col in track.columns], axis=1), how='left', on='GAME_ID')

print(features_df.shape)
features_df.head()

# load, filter (by time) and scale data
pts_features, pm_features, res_features, test_set_obs = utl.load_and_scale_data(
    input_data=features_df,
    seasons_to_keep=['2021-22', '2022-23', '2023-24'],
    training_season='2021-22',
    feature_prefixes=['ROLL_', 'ROLLDIFF_'],
    scaler_type='minmax', 
    scale_target=False
)

pts_features.head()

<a name="1.-Filter-Methods-for-Total-Points"></a>
# 1. Filter Methods for Total Points

[Return to top](#Feature-Selection)

start_time = time.time()

# filter-based feature selection
pts_selection = utl.filter_feature_selection(
    df=pts_features, 
    outcome_name='TOTAL_PTS',
    corr_threshold=0.1,
    second_stage_filter='feature_correlation',
    pairwise_corr_threshold=0.7
)

end_time = time.time()
print(f"Total time taken: {end_time - start_time:.2f} seconds")

<a name="2.-Filter-Methods-for-Plus-Minus"></a>
# 2. Filter Methods for Plus Minus

[Return to top](#Feature-Selection)

start_time = time.time()

# filter-based feature selection
pm_selection = utl.filter_feature_selection(
    df=pm_features, 
    outcome_name='PLUS_MINUS',
    corr_threshold=0.1,
    second_stage_filter='feature_correlation',
    pairwise_corr_threshold=0.7
)

end_time = time.time()
print(f"Total time taken: {end_time - start_time:.2f} seconds")

<a name="3.-Filter-Methods-for-Game-Winner"></a>
# 3. Filter Methods for Game Winner

[Return to top](#Feature-Selection)

start_time = time.time()

# filter-based feature selection
res_selection = utl.filter_feature_selection(
    df=res_features, 
    outcome_name='GAME_RESULT',
    corr_threshold=0.1,
    second_stage_filter='feature_correlation',
    pairwise_corr_threshold=0.7
)

end_time = time.time()
print(f"Total time taken: {end_time - start_time:.2f} seconds")
